Move io_devices to top, move out RomSelect
