+++ ch22-core/src/device_map.rs
@@ -3,6 +3,7 @@ use crate::devices::io_space::*;
 use crate::devices::paged_rom::*;
 use crate::devices::ram::*;
 use crate::devices::rom::Rom;
+use crate::io_devices::rom_select::RomSelect;
 use crate::word::Word;
 
 pub struct DeviceMap {
@@ -21,7 +22,7 @@ impl DeviceMap {
 
         io_space.add_device(
             &[0xfe30, 0xfe31, 0xfe32, 0xfe33],
-            Box::new(paged_rom.get_rom_select()),
+            Box::new(RomSelect::new(paged_rom.get_active_rom())),
             None,
             false,
         );
