+++ ch22-core/src/devices/paged_rom.rs
@@ -1,6 +1,5 @@
 use std::{cell::Cell, rc::Rc};
 
-use crate::devices::io_device::IODevice;
 use crate::word::Word;
 
 use super::addressable_device::*;
@@ -22,10 +21,8 @@ impl PagedRom {
         }
     }
 
-    pub fn get_rom_select(&self) -> RomSelect {
-        RomSelect {
-            active_rom: self.active_rom.clone(),
-        }
+    pub fn get_active_rom(&self) -> Rc<Cell<usize>> {
+        self.active_rom.clone()
     }
 
     pub fn load(&mut self, bank: u8, data: &[u8]) {
@@ -40,21 +37,5 @@ impl PagedRom {
 impl AddressableDevice for PagedRom {
     fn read(&mut self, address: Word, _cycles: &mut u64) -> u8 {
         self.roms[self.active_rom.get()][Into::<usize>::into(address) - self.base_address]
-    }
-}
-
-pub struct RomSelect {
-    active_rom: Rc<Cell<usize>>,
-}
-
-impl IODevice for RomSelect {
-    fn read(&mut self, _address: Word, _cycles: u64) -> u8 {
-        self.active_rom.get() as u8
-    }
-
-    fn write(&mut self, _address: Word, value: u8, _cycles: u64) -> bool {
-        self.active_rom.set(value as usize);
-
-        false
     }
 }
