+++ ch22-core/src/io_devices/static_device.rs
